Both providers now require an API key to geocode. Added this to Google and MapQuest requests.
Both utilize "results" in their JSON reply - which would incorrectly decode to the wrong pointer. Depending on provider, we forcefully decode to a different struct which fixes this.

OSM/MapQuest has a new URL, as well as JSON reply. Updated the expected JSON for decoding

Added a really nice example of leveraging both OSM/MapQuest as well as GOOGLE for batch queries